+++ src/main/java/ru/elerphore/kte/data/discount/DiscountRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DiscountRepository extends JpaRepository<DiscountEntity, Long> {
+public interface DiscountRepository extends JpaRepository<DiscountEntity, Integer> {
 }
